+++ src/lib/Scenes/SavedSearchAlert/Components/Form.tsx
@@ -1,15 +1,9 @@
 import { useFormikContext } from "formik"
 import { Input } from "lib/Components/Input/Input"
 import { InputTitle } from "lib/Components/Input/InputTitle"
-import { LegacyNativeModules } from "lib/NativeModules/LegacyNativeModules"
 import { navigate } from "lib/navigation/navigate"
-import {
-  getNotificationPermissionsStatus,
-  PushAuthorizationStatus,
-} from "lib/Scenes/MyProfile/MyProfilePushNotifications"
 import { Box, Button, Flex, Pill, Spacer, Text, Touchable } from "palette"
 import React from "react"
-import { Alert, AlertButton, Linking, Platform } from "react-native"
 import { getNamePlaceholder } from "../helpers"
 import { SavedSearchAlertFormValues, SavedSearchArtistProp } from "../SavedSearchAlertModel"
 
@@ -17,10 +11,11 @@ interface FormProps extends SavedSearchArtistProp {
   pills: string[]
   savedSearchAlertId?: string
   onDeletePress?: () => void
+  onSubmitPress?: () => void
 }
 
 export const Form: React.FC<FormProps> = (props) => {
-  const { pills, artist, savedSearchAlertId, onDeletePress } = props
+  const { pills, artist, savedSearchAlertId, onDeletePress, onSubmitPress } = props
   const {
     isSubmitting,
     values,
@@ -28,76 +23,8 @@ export const Form: React.FC<FormProps> = (props) => {
     dirty,
     handleBlur,
     handleChange,
-    handleSubmit,
   } = useFormikContext<SavedSearchAlertFormValues>()
   const namePlaceholder = getNamePlaceholder(artist.name, pills)
-
-  const requestNotificationPermissions = () => {
-    // permissions not determined: Android should never need this
-    if (Platform.OS === "ios") {
-      Alert.alert(
-        "Artsy would like to send you notifications",
-        "We need your permission to send notifications on alerts you have created.",
-        [
-          {
-            text: "Proceed",
-            onPress: () => LegacyNativeModules.ARTemporaryAPIModule.requestNotificationPermissions(),
-          },
-          {
-            text: "Cancel",
-            style: "cancel",
-          },
-        ]
-      )
-    }
-  }
-
-  const showHowToEnableNotificationInstructionAlert = () => {
-    const deviceText = Platform.select({
-      ios: "iOS",
-      android: "android",
-      default: "device",
-    })
-    const instruction = Platform.select({
-      ios: `Tap 'Artsy' and enable "Allow Notifications" for Artsy.`,
-      default: "",
-    })
-
-    const buttons: AlertButton[] = [
-      {
-        text: "Settings",
-        onPress: () => {
-          if (Platform.OS === "android") {
-            Linking.openSettings()
-          } else {
-            Linking.openURL("App-prefs:NOTIFICATIONS_ID")
-          }
-        },
-      },
-      {
-        text: "Cancel",
-        style: "cancel",
-      },
-    ]
-
-    Alert.alert(
-      "Artsy would like to send you notifications",
-      `To receive notifications for your alerts, you will need to enable them in your ${deviceText} Settings. ${instruction}`,
-      Platform.OS === "ios" ? buttons : buttons.reverse()
-    )
-  }
-
-  const onSubmitPress = async () => {
-    const notificationStatus = await getNotificationPermissionsStatus()
-
-    if (notificationStatus === PushAuthorizationStatus.Authorized) {
-      handleSubmit()
-    } else if (notificationStatus === PushAuthorizationStatus.Denied) {
-      showHowToEnableNotificationInstructionAlert()
-    } else {
-      requestNotificationPermissions()
-    }
-  }
 
   return (
     <Box>
